refactor: move mode-ids.ts to src/constants/
Removes the tools/ → modes/ import edge. mode-ids.ts is a leaf
dependency (imports nothing) used by 15 files across modes/, tools/,
and orchestration — a neutral constants/ directory respects the
dependency layering.
