+++ packages/luca-mastracode/src/modes/triage.ts
@@ -11,7 +11,7 @@ import { fileURLToPath } from 'node:url'
 import { readLucaState } from '../luca-store.js'
 import { resolveModel } from '../model-routing.js'
 
-import { MODES } from './mode-ids.js'
+import { MODES } from '../constants/mode-ids.js'
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
